[refactor] 코드 리팩토링. 컴포넌트 세분화
