--- src/components/Home/memo.txt
@@ -1,1 +1,0 @@
-공통 컴포넌트
