+++ src/components/Post/BannerSlider.tsx
@@ -50,7 +50,7 @@ function BannerSlider() {
   }, [currentIndex]);
 
   return (
-    <section className='relative w-full overflow-hidden py-6'>
+    <aside className='relative w-full overflow-hidden py-6'>
       <div
         ref={sliderRef}
         className='flex transition-transform duration-500 ease-in-out'
@@ -93,7 +93,7 @@ function BannerSlider() {
       <div className='absolute bottom-8 left-1/2 -translate-x-1/2 rounded-lg bg-[rgb(0,0,0,0.4)] px-4 text-sm text-brand-white'>
         {currentIndex + 1} / {totalSlides}
       </div>
-    </section>
+    </aside>
   );
 }
