+++ src/pages/Home.tsx
@@ -1,11 +1,8 @@
 import '../css/tailwind.css';
-import { useEffect, useRef, useState } from 'react';
-import JobPostCard from '../components/Post/JobPostCard';
+import { useState } from 'react';
 import SortDropdown from '../components/Post/SortDropdown';
-import { GotchaPostType } from '../types/gotchaPostType';
-import { useJobPost } from '../hooks/useJobPost';
 import BannerSlider from '../components/Post/BannerSlider';
-import SkeletonJobPostCard from '../components/Post/SkeletonJobPostCard';
+import JobPostList from '../components/Post/JobPostList';
 // import { supabase } from '../api/supabase/supabaseClient';
 // import { useUserStore } from '../stores/useUserStore';
 
@@ -29,68 +26,21 @@ function Home() {
     마감순: ['expiration_date', true],
   };
   const [sortLabel, setSortLabel] = useState('최신순');
-  const infinitePageSize = 5;
-  const observerRef = useRef<HTMLDivElement | null>(null);
-  const {
-    data: posts,
-    fetchNextPage,
-    hasNextPage,
-    isFetchingNextPage,
-  } = useJobPost(
-    `posts-${sortLabel}`,
-    sortLabelList[sortLabel][0],
-    sortLabelList[sortLabel][1],
-    infinitePageSize,
-  );
-
-  const files = import.meta.glob('/src/assets/company/*');
-  const companyImages = Object.keys(files);
-
-  useEffect(() => {
-    if (!observerRef.current) return;
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting && hasNextPage && !isFetchingNextPage) {
-          fetchNextPage();
-        }
-      },
-      { threshold: 1.0 },
-    );
-
-    observer.observe(observerRef.current);
-    return () => observer.disconnect();
-  }, [hasNextPage, isFetchingNextPage, fetchNextPage]);
 
   return (
     <div className='response-page-padding'>
       <BannerSlider />
-      <div className='flex justify-between'>
-        <h1 className='page-title'>채용공고</h1>
-        <SortDropdown
-          sortLabel={sortLabel}
-          setSortLabel={setSortLabel}
-          sortLabelList={sortLabelList}
-        />
-      </div>
-      <ul className='flex flex-wrap'>
-        {posts?.pages.map((page) =>
-          page.map((post: GotchaPostType, idx: number) => (
-            <JobPostCard
-              key={post.post_id}
-              post={post}
-              companyImage={companyImages[idx]}
-            />
-          )),
-        )}
-
-        {isFetchingNextPage &&
-          Array.from({ length: 5 }).map((_, index) => (
-            <SkeletonJobPostCard key={`skeleton-${index}`} />
-          ))}
-
-        <div ref={observerRef} className='h-10'></div>
-      </ul>
+      <main>
+        <header className='flex justify-between'>
+          <h1 className='page-title'>채용공고</h1>
+          <SortDropdown
+            sortLabel={sortLabel}
+            setSortLabel={setSortLabel}
+            sortLabelList={sortLabelList}
+          />
+        </header>
+        <JobPostList sortLabel={sortLabel} sortLabelList={sortLabelList} />
+      </main>
     </div>
   );
 }
